+++ src/components/_shared/PointContextMenu/PointContextMenu.module.scss
@@ -89,13 +89,6 @@
   }
 }
 
-.pointContextMenuClose {
-  position: absolute;
-  top: 5px;
-  right: 0;
-  z-index: 8;
-}
-
 .pointContextMenuHeader {
   display: grid;
   grid-template-columns: 1fr auto 1fr;
